Added projectile sub class for projectiles that use particle systems
