+++ Assets/Prefabs/Projectiles/ClusterProjectile.prefab
@@ -9,8 +9,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1655789376720695905}
+  - component: {fileID: 204064605111524863}
   - component: {fileID: 254271668005075209}
-  - component: {fileID: 1655789376720695906}
   - component: {fileID: 2576749402811358335}
   - component: {fileID: 924206450}
   m_Layer: 15
@@ -35,6 +35,25 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &204064605111524863
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1655789376720695904}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a99280fe0a6d5454abe0e1c9ddcbfcf4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  projectileVelocity: 150
+  weapon: {fileID: 7485021606706960250, guid: ba63be8833f2b5d44a70bd02c1a85a99, type: 3}
+  projectileRb2D: {fileID: 2576749402811358335}
+  targetPosition: {x: 0, y: 0, z: 0}
+  bloodSplash: {fileID: 2512025392852149761, guid: 8df12292ad0e89947a87dee41a86f181, type: 3}
+  projectileParticle: {fileID: 1344357935046343875, guid: cf6634ac9202e9a4186449d668feab00, type: 3}
+  muzzleFlashParticle: {fileID: 1855075546965395947, guid: ba84c6edddad2734a95382bc309bbaae, type: 3}
 --- !u!212 &254271668005075209
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -87,25 +106,6 @@ SpriteRenderer:
   m_WasSpriteAssigned: 0
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!114 &1655789376720695906
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1655789376720695904}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9c010d6c147971a4a95a2c58d0cc2140, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  projectileVelocity: 250
-  weapon: {fileID: 7485021606706960250, guid: ba63be8833f2b5d44a70bd02c1a85a99, type: 3}
-  projectileRb2D: {fileID: 2576749402811358335}
-  projectileParticle: {fileID: 1344357935046343875, guid: cf6634ac9202e9a4186449d668feab00, type: 3}
-  muzzleFlashParticle: {fileID: 1855075546965395947, guid: ba84c6edddad2734a95382bc309bbaae, type: 3}
-  targetPosition: {x: 0, y: 0, z: 0}
-  bloodSplash: {fileID: 2512025392852149761, guid: 8df12292ad0e89947a87dee41a86f181, type: 3}
 --- !u!50 &2576749402811358335
 Rigidbody2D:
   serializedVersion: 4
