+++ Assets/Scripts/WeaponScripts/ParticleProjectile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a99280fe0a6d5454abe0e1c9ddcbfcf4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
